refactor : MainDomain 자동차 경주 준비부분 분리
