+++ src/js/Domain/MainDomain.js
@@ -1,49 +1,21 @@
 import MatchNumber from "../state/MatchNumber.js";
-import RacingCarInfo from "../state/RacingCarInfo.js";
+import PrepareRacingDomain from "./PrepareRacingDomian.js";
+import PrepareRacingView from "../PrepareRacingView.js";
 import RunRacingRenderer from "../RunRacingRenderer.js";
 import RunRacingView from "../RunRacingView.js";
 import View from "../View.js";
 
 class MainDomain {
-  static MIN_CARNAME_SIZE = 1;
-  static MAX_CARNAME_SIZE = 5;
-  #prepareRacingView = null;
   #contestantView = null;
 
-  constructor(prepareRacingView, contestantView) {
+  constructor(contestantView) {
     this.initEventListener();
-    this.#prepareRacingView = prepareRacingView;
     this.#contestantView = contestantView;
   }
 
-  set prepareRacingView(v) {
-    if (v instanceof View) this.#prepareRacingView = v;
-    else throw `invalid vie : ${v}`;
-  }
   set contestantView(v) {
     if (v instanceof View) this.#contestantView = v;
     else throw `invalid vie : ${v}`;
-  }
-
-  testCarNameSize(carName) {
-    return (
-      MainDomain.MIN_CARNAME_SIZE <= carName.trim().length &&
-      carName.trim().length <= MainDomain.MAX_CARNAME_SIZE
-    );
-  }
-
-  submitCarNames(e) {
-    if (!e.target[1].value) {
-      return;
-    }
-    const carNames = e.target[1].value.split(",");
-    const invalidCarName = carNames.find((v) => !this.testCarNameSize(v));
-    if (invalidCarName) {
-      alert("5자 이하의 자동차 이름을 입력하세요.");
-      return;
-    }
-    RacingCarInfo.setRaceParticipateCar(carNames);
-    this.#prepareRacingView.initView();
   }
 
   submitNumberOfRaces(e) {
@@ -55,9 +27,11 @@ class MainDomain {
 
   prepareGame = (e) => {
     e.preventDefault();
-    this.#prepareRacingView.setElement(e);
     if (e.submitter.id == "car-name-btn") {
-      this.submitCarNames(e);
+      const prePareRacingDomain = new PrepareRacingDomain(
+        new PrepareRacingView()
+      );
+      prePareRacingDomain.submitCarNames(e);
     } else {
       this.submitNumberOfRaces(e);
     }
