+++ src/js/index.js
@@ -1,4 +1,3 @@
 import MainDomain from "./Domain/MainDomain.js";
-import PrepareRacingView from "./PrepareRacingView.js";
 import ContestantView from "./ContestantView.js";
-new MainDomain(new PrepareRacingView(), new ContestantView());
+new MainDomain(new ContestantView());
